+++ sncosmo/nest.py
@@ -2,7 +2,7 @@
 """Simple implementation of nested sampling routine to evaluate Bayesian
 evidence."""
 
-from __future__ import print_function
+from __future__ import print_function, division
 
 import math
 import time
@@ -207,7 +207,7 @@ def nest(loglikelihood, prior, npar, nipar, nobj=50, maxiter=10000,
 
     # Nested sampling loop.
     ndecl = 0
-    logwt_old = None
+    logwt_old = -float('inf')
     time0 = time.time()
     it = 0
     while it < maxiter and loglcalls < maxcall:
@@ -270,7 +270,7 @@ def nest(loglikelihood, prior, npar, nipar, nobj=50, maxiter=10000,
             ndecl += 1
         else:
             ndecl = 0
-        if ndecl > nobj * 2 and ndecl > it / 6:
+        if ndecl > nobj * 2 and ndecl > it // 6:
             break
         logwt_old = logwt